Firewall: send IP networks in gRPC
