+++ .sqlx/query-3c59c3ff1c0de6688906aea193a3a5d89d02a827f49c087d3769db410e95e1ec.json
@@ -1,0 +1,24 @@
+{
+  "db_name": "PostgreSQL",
+  "query": "INSERT INTO aclaliasdestinationrange (alias_id, \"start\", \"end\") VALUES ($1, $2, $3) RETURNING id",
+  "describe": {
+    "columns": [
+      {
+        "ordinal": 0,
+        "name": "id",
+        "type_info": "Int8"
+      }
+    ],
+    "parameters": {
+      "Left": [
+        "Int8",
+        "Inet",
+        "Inet"
+      ]
+    },
+    "nullable": [
+      false
+    ]
+  },
+  "hash": "3c59c3ff1c0de6688906aea193a3a5d89d02a827f49c087d3769db410e95e1ec"
+}
