--- .sqlx/query-c67f95783b769e6caf3da520a582960f4ffef65dae743497343b653960a2ed19.json
@@ -1,16 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "INSERT INTO aclruledestinationrange (rule_id, \"start\", \"end\") VALUES ($1, $2, $3)",
-  "describe": {
-    "columns": [],
-    "parameters": {
-      "Left": [
-        "Int8",
-        "Inet",
-        "Inet"
-      ]
-    },
-    "nullable": []
-  },
-  "hash": "c67f95783b769e6caf3da520a582960f4ffef65dae743497343b653960a2ed19"
-}
