+++ src/enterprise/db/models/acl.rs
@@ -1,7 +1,7 @@
 use std::{
     collections::HashSet,
     fmt,
-    net::{IpAddr, Ipv4Addr},
+    net::IpAddr,
     ops::{Bound, RangeInclusive},
 };
 
@@ -78,16 +78,19 @@ impl fmt::Display for PortRange {
 }
 
 impl PortRange {
+    #[must_use]
     pub fn new(start: u16, end: u16) -> Self {
         Self(start..=end)
     }
 
-    // Returns first port in range
+    /// Returns first port in range.
+    #[must_use]
     pub fn first_port(&self) -> u16 {
         *self.0.start()
     }
 
-    // Returns last port in range
+    /// Returns last port in range.
+    #[must_use]
     pub fn last_port(&self) -> u16 {
         *self.0.end()
     }
@@ -173,7 +176,7 @@ pub struct AclRuleInfo<I = NoId> {
 }
 
 impl<I> AclRuleInfo<I> {
-    /// Constructs a [`String`] of comma-separated addresses and address ranges
+    /// Constructs a [`String`] of comma-separated addresses and address ranges.
     pub(crate) fn format_destination(&self) -> String {
         // process single addresses
         let addrs = match &self.destination {
@@ -198,19 +201,13 @@ impl<I> AclRuleInfo<I> {
         }
     }
 
-    /// Constructs a [`String`] of comma-separated ports and port ranges
+    /// Constructs a [`String`] of comma-separated ports and port ranges.
     pub(crate) fn format_ports(&self) -> String {
-        if self.ports.is_empty() {
-            String::new()
-        } else {
-            let ports = self
-                .ports
-                .iter()
-                .map(|r| r.to_string() + ", ")
-                .collect::<String>();
-            // trim the last last ', '
-            ports[..ports.len() - 2].to_string()
-        }
+        self.ports
+            .iter()
+            .map(ToString::to_string)
+            .collect::<Vec<_>>()
+            .join(", ")
     }
 }
 
@@ -371,8 +368,9 @@ impl AclRule {
     /// State handling:
     ///
     /// - For rules in `RuleState::Applied` state (rules that are currently active):
-    ///   1. Any existing modifications of this rule are deleted
-    ///   2. A copy of the rule is created with `RuleState::Deleted` state and the original rule as parent
+    ///   1. Any existing modifications of this rule are deleted.
+    ///   2. A copy of the rule is created with `RuleState::Deleted` state and the original rule as
+    ///      parent.
     ///
     /// This preserves the original rule while tracking the deletion.
     ///
@@ -426,8 +424,8 @@ impl AclRule {
             _ => {
                 // delete the not-yet applied modification itself
                 debug!(
-                    "Rule {} is a modification to rule {:?} - updating the modification",
-                    id, existing_rule.parent_id,
+                    "Rule {id} is a modification to rule {:?} - updating the modification",
+                    existing_rule.parent_id,
                 );
                 // delete related objects
                 existing_rule
@@ -437,7 +435,7 @@ impl AclRule {
                 // delete the rule
                 existing_rule.delete(&mut *transaction).await?;
             }
-        };
+        }
 
         transaction.commit().await?;
         info!("Rule {id} succesfully deleted or marked for deletion");
@@ -471,7 +469,8 @@ impl AclRule {
         let affected_locations: Vec<WireguardNetwork<Id>> =
             affected_locations.into_iter().collect();
         debug!(
-            "{} locations affected by applied ACL rules. Sending gateway firewall update events for each location",
+            "{} locations affected by applied ACL rules. Sending gateway firewall update events \
+            for each location",
             affected_locations.len()
         );
 
@@ -485,7 +484,10 @@ impl AclRule {
                     ));
                 }
                 None => {
-                    debug!("No firewall config generated for location {location}. Not sending a gateway event")
+                    debug!(
+                        "No firewall config generated for location {location}. Not sending a \
+                        gateway event"
+                    );
                 }
             }
         }
@@ -508,20 +510,19 @@ pub fn parse_destination(destination: &str) -> Result<ParsedDestination, AclErro
     debug!("Parsing destination string: {destination}");
     let destination: String = destination.chars().filter(|c| !c.is_whitespace()).collect();
     let mut result = ParsedDestination::default();
-    if destination.is_empty() {
-        return Ok(result);
-    }
-    for v in destination.split(',') {
-        match v.split('-').collect::<Vec<_>>() {
-            l if l.len() == 1 => result.addrs.push(l[0].parse::<IpNetwork>()?),
-            l if l.len() == 2 => result
-                .ranges
-                .push((l[0].parse::<IpAddr>()?, l[1].parse::<IpAddr>()?)),
-            _ => {
-                error!("Failed to parse destination string: \"{destination}\"");
-                Err(IpNetworkError::InvalidAddr(destination.clone()))?;
+    if !destination.is_empty() {
+        for v in destination.split(',') {
+            match v.split('-').collect::<Vec<_>>() {
+                l if l.len() == 1 => result.addrs.push(l[0].parse::<IpNetwork>()?),
+                l if l.len() == 2 => result
+                    .ranges
+                    .push((l[0].parse::<IpAddr>()?, l[1].parse::<IpAddr>()?)),
+                _ => {
+                    error!("Failed to parse destination string: \"{destination}\"");
+                    Err(IpNetworkError::InvalidAddr(destination.clone()))?;
+                }
             }
-        };
+        }
     }
 
     debug!("Parsed destination: {result:?}");
@@ -543,18 +544,18 @@ pub fn parse_ports(ports: &str) -> Result<Vec<PortRange>, AclError> {
             match v.split('-').collect::<Vec<_>>() {
                 l if l.len() == 1 => {
                     let start = l[0].parse::<u16>()?;
-                    result.push(PortRange::new(start, start))
+                    result.push(PortRange::new(start, start));
                 }
                 l if l.len() == 2 => {
                     let start = l[0].parse::<u16>()?;
                     let end = l[1].parse::<u16>()?;
-                    result.push(PortRange::new(start, end))
+                    result.push(PortRange::new(start, end));
                 }
                 _ => {
                     error!("Failed to parse ports string: \"{ports}\"");
                     return Err(AclError::InvalidPortsFormat(ports.to_string()));
                 }
-            };
+            }
         }
     }
 
@@ -563,7 +564,7 @@ pub fn parse_ports(ports: &str) -> Result<Vec<PortRange>, AclError> {
 }
 
 /// Maps [`sqlx::Error`] to [`AclError`] while checking for [`ErrorKind::ForeignKeyViolation`].
-fn map_relation_error(err: SqlxError, class: &str, id: &Id) -> AclError {
+fn map_relation_error(err: SqlxError, class: &str, id: Id) -> AclError {
     if let SqlxError::Database(dberror) = &err {
         if dberror.kind() == ErrorKind::ForeignKeyViolation {
             error!("Failed to create ACL related object, foreign key violation: {class}({id}): {dberror}");
@@ -593,7 +594,7 @@ impl AclRule<Id> {
             };
             obj.save(&mut *transaction)
                 .await
-                .map_err(|err| map_relation_error(err, "WireguardNetwork", network_id))?;
+                .map_err(|err| map_relation_error(err, "WireguardNetwork", *network_id))?;
         }
 
         // allowed users
@@ -607,7 +608,7 @@ impl AclRule<Id> {
             };
             obj.save(&mut *transaction)
                 .await
-                .map_err(|err| map_relation_error(err, "User", user_id))?;
+                .map_err(|err| map_relation_error(err, "User", *user_id))?;
         }
 
         // denied users
@@ -621,7 +622,7 @@ impl AclRule<Id> {
             };
             obj.save(&mut *transaction)
                 .await
-                .map_err(|err| map_relation_error(err, "User", user_id))?;
+                .map_err(|err| map_relation_error(err, "User", *user_id))?;
         }
 
         // allowed groups
@@ -635,7 +636,7 @@ impl AclRule<Id> {
             };
             obj.save(&mut *transaction)
                 .await
-                .map_err(|err| map_relation_error(err, "Group", group_id))?;
+                .map_err(|err| map_relation_error(err, "Group", *group_id))?;
         }
 
         // denied groups
@@ -649,7 +650,7 @@ impl AclRule<Id> {
             };
             obj.save(&mut *transaction)
                 .await
-                .map_err(|err| map_relation_error(err, "Group", group_id))?;
+                .map_err(|err| map_relation_error(err, "Group", *group_id))?;
         }
 
         // save related aliases
@@ -668,7 +669,7 @@ impl AclRule<Id> {
             return Err(AclError::CannotUseModifiedAliasInRuleError(
                 invalid_alias_ids,
             ));
-        };
+        }
         for alias_id in &api_rule.aliases {
             let obj = AclRuleAlias {
                 id: NoId,
@@ -677,7 +678,7 @@ impl AclRule<Id> {
             };
             obj.save(&mut *transaction)
                 .await
-                .map_err(|err| map_relation_error(err, "AclAlias", alias_id))?;
+                .map_err(|err| map_relation_error(err, "AclAlias", *alias_id))?;
         }
 
         // allowed devices
@@ -691,7 +692,7 @@ impl AclRule<Id> {
             };
             obj.save(&mut *transaction)
                 .await
-                .map_err(|err| map_relation_error(err, "Device", device_id))?;
+                .map_err(|err| map_relation_error(err, "Device", *device_id))?;
         }
 
         // denied devices
@@ -705,7 +706,7 @@ impl AclRule<Id> {
             };
             obj.save(&mut *transaction)
                 .await
-                .map_err(|err| map_relation_error(err, "Device", device_id))?;
+                .map_err(|err| map_relation_error(err, "Device", *device_id))?;
         }
 
         // destination
@@ -834,7 +835,7 @@ impl AclRule<Id> {
     /// - clears rule's `parent_id`.
     /// - deletes it's parent rule
     ///
-    /// If current state is ['RuleState::Deleted'] it removes the parent rule and the rule itself.
+    /// If current state is [`RuleState::Deleted`] it removes the parent rule and the rule itself.
     ///
     /// # Errors
     ///
@@ -922,7 +923,8 @@ impl AclRule<Id> {
     {
         query_as!(
             AclAlias,
-            "SELECT a.id, parent_id, name, kind \"kind: AliasKind\",state \"state: AliasState\", destination, ports, protocols \
+            "SELECT a.id, parent_id, name, kind \"kind: AliasKind\",state \"state: AliasState\", \
+            destination, ports, protocols \
             FROM aclrulealias r \
             JOIN aclalias a \
             ON a.id = r.alias_id \
@@ -942,9 +944,10 @@ impl AclRule<Id> {
     where
         E: PgExecutor<'e>,
     {
-        Ok(match allowed {
-            true => self.get_allowed_users(executor).await?,
-            false => self.get_denied_users(executor).await?,
+        Ok(if allowed {
+            self.get_allowed_users(executor).await?
+        } else {
+            self.get_denied_users(executor).await?
         })
     }
 
@@ -958,16 +961,16 @@ impl AclRule<Id> {
     {
         query_as!(
             User,
-            "SELECT u.id, username, password_hash, last_name, first_name, email, \
-                phone, mfa_enabled, totp_enabled, totp_secret, \
-                email_mfa_enabled, email_mfa_secret, \
-                mfa_method \"mfa_method: _\", recovery_codes, is_active, openid_sub, from_ldap, ldap_pass_randomized, ldap_rdn \
-                FROM aclruleuser r \
-                JOIN \"user\" u \
-                ON u.id = r.user_id \
-                WHERE r.rule_id = $1 \
-                AND r.allow \
-                AND u.is_active = true",
+            "SELECT u.id, username, password_hash, last_name, first_name, email, phone, \
+            mfa_enabled, totp_enabled, totp_secret, email_mfa_enabled, email_mfa_secret, \
+            mfa_method \"mfa_method: _\", recovery_codes, is_active, openid_sub, from_ldap, \
+            ldap_pass_randomized, ldap_rdn \
+            FROM aclruleuser r \
+            JOIN \"user\" u \
+            ON u.id = r.user_id \
+            WHERE r.rule_id = $1 \
+            AND r.allow \
+            AND u.is_active = true",
             self.id,
         )
         .fetch_all(executor)
@@ -984,16 +987,16 @@ impl AclRule<Id> {
     {
         query_as!(
             User,
-            "SELECT u.id, username, password_hash, last_name, first_name, email, \
-                phone, mfa_enabled, totp_enabled, totp_secret, \
-                email_mfa_enabled, email_mfa_secret, \
-                mfa_method \"mfa_method: _\", recovery_codes, is_active, openid_sub, from_ldap, ldap_pass_randomized, ldap_rdn \
-                FROM aclruleuser r \
-                JOIN \"user\" u \
-                ON u.id = r.user_id \
-                WHERE r.rule_id = $1 \
-                AND NOT r.allow \
-                AND u.is_active = true",
+            "SELECT u.id, username, password_hash, last_name, first_name, email, phone, \
+            mfa_enabled, totp_enabled, totp_secret, email_mfa_enabled, email_mfa_secret, \
+            mfa_method \"mfa_method: _\", recovery_codes, is_active, openid_sub, from_ldap, \
+            ldap_pass_randomized, ldap_rdn \
+            FROM aclruleuser r \
+            JOIN \"user\" u \
+            ON u.id = r.user_id \
+            WHERE r.rule_id = $1 \
+            AND NOT r.allow \
+            AND u.is_active = true",
             self.id,
         )
         .fetch_all(executor)
@@ -1033,9 +1036,10 @@ impl AclRule<Id> {
     where
         E: PgExecutor<'e>,
     {
-        match allowed {
-            true => self.get_allowed_network_devices(executor).await,
-            false => self.get_denied_network_devices(executor).await,
+        if allowed {
+            self.get_allowed_network_devices(executor).await
+        } else {
+            self.get_denied_network_devices(executor).await
         }
     }
 
@@ -1047,18 +1051,18 @@ impl AclRule<Id> {
         E: PgExecutor<'e>,
     {
         query_as!(
-                Device,
-                "SELECT d.id, name, wireguard_pubkey, user_id, created, description, device_type \"device_type: DeviceType\", \
-                    configured \
-                    FROM aclruledevice r \
-                    JOIN device d \
-                    ON d.id = r.device_id \
-                    WHERE r.rule_id = $1 \
-                    AND r.allow = true AND d.configured = true",
-                self.id,
-            )
-                .fetch_all(executor)
-            .await
+            Device,
+            "SELECT d.id, name, wireguard_pubkey, user_id, created, description, \
+            device_type \"device_type: DeviceType\", configured \
+            FROM aclruledevice r \
+            JOIN device d \
+            ON d.id = r.device_id \
+            WHERE r.rule_id = $1 \
+            AND r.allow = true AND d.configured = true",
+            self.id,
+        )
+        .fetch_all(executor)
+        .await
     }
 
     pub(crate) async fn get_denied_network_devices<'e, E>(
@@ -1069,18 +1073,18 @@ impl AclRule<Id> {
         E: PgExecutor<'e>,
     {
         query_as!(
-                Device,
-                "SELECT d.id, name, wireguard_pubkey, user_id, created, description, device_type \"device_type: DeviceType\", \
-                    configured \
-                    FROM aclruledevice r \
-                    JOIN device d \
-                    ON d.id = r.device_id \
-                    WHERE r.rule_id = $1 \
-                    AND r.allow = false AND d.configured = true",
-                self.id,
-            )
-                .fetch_all(executor)
-            .await
+            Device,
+            "SELECT d.id, name, wireguard_pubkey, user_id, created, description, \
+            device_type \"device_type: DeviceType\", configured \
+            FROM aclruledevice r \
+            JOIN device d \
+            ON d.id = r.device_id \
+            WHERE r.rule_id = $1 \
+            AND r.allow = false AND d.configured = true",
+            self.id,
+        )
+        .fetch_all(executor)
+        .await
     }
 
     /// Returns all [`AclRuleDestinationRanges`]es the rule applies to
@@ -1166,7 +1170,8 @@ impl AclRuleInfo<Id> {
                 "SELECT id, username, password_hash, last_name, first_name, email, \
                 phone, mfa_enabled, totp_enabled, totp_secret, \
                 email_mfa_enabled, email_mfa_secret, \
-                mfa_method \"mfa_method: _\", recovery_codes, is_active, openid_sub, from_ldap, ldap_pass_randomized, ldap_rdn \
+                mfa_method \"mfa_method: _\", recovery_codes, is_active, openid_sub, from_ldap, \
+                ldap_pass_randomized, ldap_rdn \
                 FROM \"user\" \
                 WHERE is_active = true"
             )
@@ -1185,14 +1190,13 @@ impl AclRuleInfo<Id> {
         // fetch all active members of allowed groups
         let allowed_groups_users: Vec<User<Id>> = query_as!(
             User,
-            "SELECT id, username, password_hash, last_name, first_name, email, \
-                phone, mfa_enabled, totp_enabled, totp_secret, \
-                email_mfa_enabled, email_mfa_secret, \
-                mfa_method \"mfa_method: _\", recovery_codes, is_active, openid_sub, \
-                from_ldap, ldap_pass_randomized, ldap_rdn \
-                FROM \"user\" u \
-                JOIN group_user gu ON u.id=gu.user_id \
-                WHERE u.is_active=true AND gu.group_id=ANY($1)",
+            "SELECT id, username, password_hash, last_name, first_name, email, phone, mfa_enabled, \
+            totp_enabled, totp_secret, email_mfa_enabled, email_mfa_secret, \
+            mfa_method \"mfa_method: _\", recovery_codes, is_active, openid_sub, \
+            from_ldap, ldap_pass_randomized, ldap_rdn \
+            FROM \"user\" u \
+            JOIN group_user gu ON u.id=gu.user_id \
+            WHERE u.is_active=true AND gu.group_id=ANY($1)",
             &allowed_group_ids
         )
         .fetch_all(executor)
@@ -1227,7 +1231,8 @@ impl AclRuleInfo<Id> {
                 "SELECT id, username, password_hash, last_name, first_name, email, \
                 phone, mfa_enabled, totp_enabled, totp_secret, \
                 email_mfa_enabled, email_mfa_secret, \
-                mfa_method \"mfa_method: _\", recovery_codes, is_active, openid_sub, from_ldap, ldap_pass_randomized, ldap_rdn \
+                mfa_method \"mfa_method: _\", recovery_codes, is_active, openid_sub, from_ldap, \
+                ldap_pass_randomized, ldap_rdn \
                 FROM \"user\" \
                 WHERE is_active = true"
             )
@@ -1268,7 +1273,7 @@ impl AclRuleInfo<Id> {
     }
 
     /// Returns the list of explicitly configured allowed network devices or
-    /// a list of all devices if 'allow_all_network_devices' flag is enabled
+    /// a list of all devices if 'allow_all_network_devices' flag is enabled.
     pub(crate) async fn get_all_allowed_devices<'e, E: sqlx::PgExecutor<'e>>(
         &self,
         executor: E,
@@ -1280,26 +1285,27 @@ impl AclRuleInfo<Id> {
         );
         // return all active devices if `allow_all_network_devices` flag is enabled
         if self.allow_all_network_devices {
-            return query_as!(
+            query_as!(
                 Device,
-                "SELECT d.id, name, wireguard_pubkey, user_id, created, description, device_type \"device_type: DeviceType\", \
-                    configured \
-                    FROM device d \
-                    JOIN wireguard_network_device wnd \
-                    ON d.id = wnd.device_id \
-                    WHERE device_type = 'network'::device_type AND configured = true AND wireguard_network_id = $1",
+                "SELECT d.id, name, wireguard_pubkey, user_id, created, description, \
+                device_type \"device_type: DeviceType\", configured \
+                FROM device d \
+                JOIN wireguard_network_device wnd \
+                ON d.id = wnd.device_id \
+                WHERE device_type = 'network'::device_type AND configured = true AND \
+                wireguard_network_id = $1",
                 location_id
             )
-                .fetch_all(executor)
-            .await;
+            .fetch_all(executor)
+            .await
+        } else {
+            // return explicitly configured allowed devices otherwise
+            Ok(self.allowed_devices.clone())
         }
-
-        // return explicitly configured allowed devices otherwise
-        Ok(self.allowed_devices.clone())
     }
 
     /// Returns the list of explicitly configured denied network devices or
-    /// a list of all devices if 'deny_all_network_devices' flag is enabled
+    /// a list of all devices if 'deny_all_network_devices' flag is enabled.
     pub(crate) async fn get_all_denied_devices<'e, E: sqlx::PgExecutor<'e>>(
         &self,
         executor: E,
@@ -1311,22 +1317,23 @@ impl AclRuleInfo<Id> {
         );
         // return all active devices if `allow_all_network_devices` flag is enabled
         if self.deny_all_network_devices {
-            return query_as!(
+            query_as!(
                 Device,
-                "SELECT d.id, name, wireguard_pubkey, user_id, created, description, device_type \"device_type: DeviceType\", \
-                    configured \
-                    FROM device d \
-                    JOIN wireguard_network_device wnd \
-                    ON d.id = wnd.device_id \
-                    WHERE device_type = 'network'::device_type AND configured = true AND wireguard_network_id = $1",
+                "SELECT d.id, name, wireguard_pubkey, user_id, created, description, \
+                device_type \"device_type: DeviceType\", configured \
+                FROM device d \
+                JOIN wireguard_network_device wnd \
+                ON d.id = wnd.device_id \
+                WHERE device_type = 'network'::device_type AND configured = true AND \
+                wireguard_network_id = $1",
                 location_id
             )
-                .fetch_all(executor)
-            .await;
+            .fetch_all(executor)
+            .await
+        } else {
+            // return explicitly configured denied devices otherwise
+            Ok(self.denied_devices.clone())
         }
-
-        // return explicitly configured denied devices otherwise
-        Ok(self.denied_devices.clone())
     }
 }
 
@@ -1374,17 +1381,11 @@ impl<I> AclAliasInfo<I> {
 
     /// Constructs a [`String`] of comma-separated ports and port ranges
     pub fn format_ports(&self) -> String {
-        if self.ports.is_empty() {
-            String::new()
-        } else {
-            let ports = self
-                .ports
-                .iter()
-                .map(|r| r.to_string() + ", ")
-                .collect::<String>();
-            // trim the last last ', '
-            ports[..ports.len() - 2].to_string()
-        }
+        self.ports
+            .iter()
+            .map(ToString::to_string)
+            .collect::<Vec<_>>()
+            .join(", ")
     }
 }
 
@@ -1605,8 +1606,8 @@ impl AclAlias {
             .await?;
         let removed_modifications = result.rows_affected();
         if removed_modifications > 0 {
-            debug!("Removed {removed_modifications} old modifications of alias {id}",);
-        };
+            debug!("Removed {removed_modifications} old modifications of alias {id}");
+        }
 
         // delete related objects
         Self::delete_related_objects(&mut transaction, id).await?;
@@ -1659,7 +1660,8 @@ impl AclAlias {
         let affected_locations: Vec<WireguardNetwork<Id>> =
             affected_locations.into_iter().collect();
         debug!(
-            "{} locations affected by applied ACL aliases. Sending gateway firewall update events for each location",
+            "{} locations affected by applied ACL aliases. Sending gateway firewall update events \
+            for each location",
             affected_locations.len()
         );
 
@@ -1673,7 +1675,10 @@ impl AclAlias {
                     ));
                 }
                 None => {
-                    debug!("No firewall config generated for location {location}. Not sending a gateway event")
+                    debug!(
+                        "No firewall config generated for location {location}. Not sending a \
+                        gateway event"
+                    );
                 }
             }
         }
@@ -1683,8 +1688,8 @@ impl AclAlias {
     }
 }
 
-impl<I: std::fmt::Debug> AclAlias<I> {
-    /// Creates relation objects for given [`AclAlias`] based on [`AclAliasInfo`] object
+impl<I: fmt::Debug> AclAlias<I> {
+    /// Creates relation objects for a given [`AclAlias`] based on [`AclAliasInfo`] object.
     async fn create_related_objects(
         transaction: &mut PgConnection,
         alias_id: Id,
@@ -1707,7 +1712,7 @@ impl<I: std::fmt::Debug> AclAlias<I> {
         Ok(())
     }
 
-    /// Deletes relation objects for given [`AclAlias`]
+    /// Deletes relation objects for a given [`AclAlias`].
     async fn delete_related_objects(
         transaction: &mut PgConnection,
         alias_id: Id,
@@ -1844,38 +1849,38 @@ impl AclAlias<Id> {
 #[derive(Clone, Debug, Model, PartialEq)]
 pub struct AclRuleNetwork<I = NoId> {
     pub id: I,
-    pub rule_id: i64,
-    pub network_id: i64,
+    pub rule_id: Id,
+    pub network_id: Id,
 }
 
 #[derive(Clone, Debug, Model, PartialEq)]
 pub struct AclRuleUser<I = NoId> {
     pub id: I,
-    pub rule_id: i64,
-    pub user_id: i64,
+    pub rule_id: Id,
+    pub user_id: Id,
     pub allow: bool,
 }
 
 #[derive(Clone, Debug, Model, PartialEq)]
 pub struct AclRuleGroup<I = NoId> {
     pub id: I,
-    pub rule_id: i64,
-    pub group_id: i64,
+    pub rule_id: Id,
+    pub group_id: Id,
     pub allow: bool,
 }
 
 #[derive(Clone, Debug, Model, PartialEq)]
 pub struct AclRuleAlias<I = NoId> {
     pub id: I,
-    pub rule_id: i64,
-    pub alias_id: i64,
+    pub rule_id: Id,
+    pub alias_id: Id,
 }
 
 #[derive(Clone, Debug, Model, PartialEq)]
 pub struct AclRuleDevice<I = NoId> {
     pub id: I,
-    pub rule_id: i64,
-    pub device_id: i64,
+    pub rule_id: Id,
+    pub device_id: Id,
     pub allow: bool,
 }
 
@@ -1887,34 +1892,39 @@ pub struct AclRuleDestinationRange<I = NoId> {
     pub end: IpAddr,
 }
 
-impl Default for AclRuleDestinationRange<Id> {
-    fn default() -> Self {
-        Self {
-            id: Id::default(),
-            rule_id: Id::default(),
-            start: IpAddr::V4(Ipv4Addr::UNSPECIFIED),
-            end: IpAddr::V4(Ipv4Addr::UNSPECIFIED),
-        }
+impl<I> AclRuleDestinationRange<I> {
+    pub(crate) fn fits_in_network(&self, ipnet: &IpNetwork) -> bool {
+        ipnet.contains(self.start) && ipnet.contains(self.end)
     }
 }
 
 impl AclRuleDestinationRange<NoId> {
-    pub async fn save<'e, E>(&self, executor: E) -> Result<(), SqlxError>
+    pub async fn save<'e, E>(self, executor: E) -> Result<AclRuleDestinationRange<Id>, SqlxError>
     where
         E: PgExecutor<'e>,
     {
-        query!(
-            "INSERT INTO aclruledestinationrange \
-            (rule_id, \"start\", \"end\") \
-            VALUES ($1, $2, $3)",
+        let id = query_scalar!(
+            "INSERT INTO aclruledestinationrange (rule_id, \"start\", \"end\") \
+            VALUES ($1, $2, $3) RETURNING id",
             self.rule_id,
             IpNetwork::from(self.start),
             IpNetwork::from(self.end),
         )
-        .execute(executor)
+        .fetch_one(executor)
         .await?;
 
-        Ok(())
+        Ok(AclRuleDestinationRange {
+            id,
+            rule_id: self.rule_id,
+            start: self.start,
+            end: self.end,
+        })
+    }
+}
+
+impl<I> From<&AclRuleDestinationRange<I>> for RangeInclusive<IpAddr> {
+    fn from(value: &AclRuleDestinationRange<I>) -> Self {
+        value.start..=value.end
     }
 }
 
@@ -1926,743 +1936,41 @@ pub struct AclAliasDestinationRange<I = NoId> {
     pub end: IpAddr,
 }
 
-impl Default for AclAliasDestinationRange<Id> {
-    fn default() -> Self {
-        Self {
-            id: Id::default(),
-            alias_id: Id::default(),
-            start: IpAddr::V4(Ipv4Addr::UNSPECIFIED),
-            end: IpAddr::V4(Ipv4Addr::UNSPECIFIED),
-        }
+impl<I> AclAliasDestinationRange<I> {
+    pub(crate) fn fits_in_network(&self, ipnet: &IpNetwork) -> bool {
+        ipnet.contains(self.start) && ipnet.contains(self.end)
     }
 }
 
 impl AclAliasDestinationRange<NoId> {
-    pub async fn save<'e, E>(&self, executor: E) -> Result<(), SqlxError>
+    pub async fn save<'e, E>(self, executor: E) -> Result<AclAliasDestinationRange<Id>, SqlxError>
     where
         E: PgExecutor<'e>,
     {
-        query!(
-            "INSERT INTO aclaliasdestinationrange \
-            (alias_id, \"start\", \"end\") \
-            VALUES ($1, $2, $3)",
+        let id = query_scalar!(
+            "INSERT INTO aclaliasdestinationrange (alias_id, \"start\", \"end\") \
+            VALUES ($1, $2, $3) RETURNING id",
             self.alias_id,
             IpNetwork::from(self.start),
             IpNetwork::from(self.end),
         )
-        .execute(executor)
+        .fetch_one(executor)
         .await?;
 
-        Ok(())
+        Ok(AclAliasDestinationRange {
+            id,
+            alias_id: self.alias_id,
+            start: self.start,
+            end: self.end,
+        })
+    }
+}
+
+impl<I> From<&AclAliasDestinationRange<I>> for RangeInclusive<IpAddr> {
+    fn from(value: &AclAliasDestinationRange<I>) -> Self {
+        value.start..=value.end
     }
 }
 
 #[cfg(test)]
-mod test {
-    use std::ops::Bound;
-
-    use rand::{thread_rng, Rng};
-    use sqlx::postgres::{PgConnectOptions, PgPoolOptions};
-
-    use super::*;
-    use crate::{db::setup_pool, handlers::wireguard::parse_address_list};
-
-    #[sqlx::test]
-    async fn test_alias(_: PgPoolOptions, options: PgConnectOptions) {
-        let pool = setup_pool(options).await;
-
-        let destination = parse_address_list("10.0.0.1, 10.1.0.0/16");
-        let ports = vec![
-            PgRange {
-                start: Bound::Included(10),
-                end: Bound::Excluded(21),
-            },
-            PgRange {
-                start: Bound::Included(100),
-                end: Bound::Excluded(201),
-            },
-        ];
-        let alias = AclAlias::new(
-            "alias",
-            AliasState::Applied,
-            AliasKind::Destination,
-            destination.clone(),
-            ports.clone(),
-            vec![20, 30],
-        )
-        .save(&pool)
-        .await
-        .unwrap();
-
-        assert_eq!(alias.id, 1);
-
-        let retrieved = AclAlias::find_by_id(&pool, 1).await.unwrap().unwrap();
-
-        assert_eq!(retrieved.id, 1);
-        assert_eq!(retrieved.destination, destination);
-        assert_eq!(retrieved.ports, ports);
-    }
-
-    #[sqlx::test]
-    async fn test_allow_conflicting_sources(_: PgPoolOptions, options: PgConnectOptions) {
-        let pool = setup_pool(options).await;
-
-        // create the rule
-        let rule = AclRule {
-            id: NoId,
-            parent_id: Default::default(),
-            state: Default::default(),
-            name: "rule".to_string(),
-            enabled: true,
-            allow_all_users: false,
-            deny_all_users: false,
-            allow_all_network_devices: false,
-            deny_all_network_devices: false,
-            all_networks: false,
-            destination: Vec::new(),
-            ports: Vec::new(),
-            protocols: Vec::new(),
-            expires: None,
-        }
-        .save(&pool)
-        .await
-        .unwrap();
-
-        // user
-        let user = User::new("user1", None, "", "", "u1@mail.com", None)
-            .save(&pool)
-            .await
-            .unwrap();
-        let _ = AclRuleUser {
-            id: NoId,
-            rule_id: rule.id,
-            user_id: user.id,
-            allow: true,
-        }
-        .save(&pool)
-        .await
-        .unwrap();
-        let result = AclRuleUser {
-            id: NoId,
-            rule_id: rule.id,
-            user_id: user.id,
-            allow: false,
-        }
-        .save(&pool)
-        .await;
-        assert!(result.is_ok());
-
-        // group
-        let group = Group::new("group1").save(&pool).await.unwrap();
-        let _ = AclRuleGroup {
-            id: NoId,
-            rule_id: rule.id,
-            group_id: group.id,
-            allow: true,
-        }
-        .save(&pool)
-        .await
-        .unwrap();
-        let result = AclRuleGroup {
-            id: NoId,
-            rule_id: rule.id,
-            group_id: group.id,
-            allow: false,
-        }
-        .save(&pool)
-        .await;
-        assert!(result.is_ok());
-
-        // device
-        let device = Device::new(
-            "device1".to_string(),
-            String::new(),
-            1,
-            DeviceType::Network,
-            None,
-            true,
-        )
-        .save(&pool)
-        .await
-        .unwrap();
-        let _ = AclRuleDevice {
-            id: NoId,
-            rule_id: rule.id,
-            device_id: device.id,
-            allow: true,
-        }
-        .save(&pool)
-        .await
-        .unwrap();
-        let result = AclRuleDevice {
-            id: NoId,
-            rule_id: rule.id,
-            device_id: device.id,
-            allow: false,
-        }
-        .save(&pool)
-        .await;
-        assert!(result.is_ok());
-    }
-
-    #[sqlx::test]
-    async fn test_rule_relations(_: PgPoolOptions, options: PgConnectOptions) {
-        let pool = setup_pool(options).await;
-
-        // create the rule
-        let mut rule = AclRule {
-            id: NoId,
-            parent_id: Default::default(),
-            state: Default::default(),
-            name: "rule".to_string(),
-            enabled: true,
-            allow_all_users: false,
-            deny_all_users: false,
-            allow_all_network_devices: false,
-            deny_all_network_devices: false,
-            all_networks: false,
-            destination: Vec::new(),
-            ports: Vec::new(),
-            protocols: Vec::new(),
-            expires: None,
-        }
-        .save(&pool)
-        .await
-        .unwrap();
-
-        // create 2 networks
-        let network1 = WireguardNetwork::new(
-            "network1".to_string(),
-            Vec::new(),
-            1000,
-            "endpoint1".to_string(),
-            None,
-            Vec::new(),
-            false,
-            100,
-            100,
-            false,
-            false,
-        )
-        .unwrap()
-        .save(&pool)
-        .await
-        .unwrap();
-        let _network2 = WireguardNetwork::new(
-            "network2".to_string(),
-            Vec::new(),
-            2000,
-            "endpoint2".to_string(),
-            None,
-            Vec::new(),
-            false,
-            200,
-            200,
-            false,
-            false,
-        )
-        .unwrap()
-        .save(&pool)
-        .await
-        .unwrap();
-
-        // rule only applied to network1
-        let _rn = AclRuleNetwork {
-            id: NoId,
-            rule_id: rule.id,
-            network_id: network1.id,
-        }
-        .save(&pool)
-        .await
-        .unwrap();
-
-        // create 2 users
-        let mut user1 = User::new("user1", None, "", "", "u1@mail.com", None)
-            .save(&pool)
-            .await
-            .unwrap();
-        let user2 = User::new("user2", None, "", "", "u2@mail.com", None)
-            .save(&pool)
-            .await
-            .unwrap();
-
-        // user1 allowed
-        let _ru1 = AclRuleUser {
-            id: NoId,
-            rule_id: rule.id,
-            user_id: user1.id,
-            allow: true,
-        }
-        .save(&pool)
-        .await
-        .unwrap();
-
-        // user2 denied
-        let mut ru2 = AclRuleUser {
-            id: NoId,
-            rule_id: rule.id,
-            user_id: user2.id,
-            allow: false,
-        }
-        .save(&pool)
-        .await
-        .unwrap();
-
-        // create 2 grups
-        let group1 = Group::new("group1").save(&pool).await.unwrap();
-        let group2 = Group::new("group2").save(&pool).await.unwrap();
-
-        // group1 allowed
-        let _rg = AclRuleGroup {
-            id: NoId,
-            rule_id: rule.id,
-            group_id: group1.id,
-            allow: true,
-        }
-        .save(&pool)
-        .await
-        .unwrap();
-
-        // group2 denied
-        let _rg = AclRuleGroup {
-            id: NoId,
-            rule_id: rule.id,
-            group_id: group2.id,
-            allow: false,
-        }
-        .save(&pool)
-        .await
-        .unwrap();
-
-        // create 2 devices
-        let device1 = Device::new(
-            "device1".to_string(),
-            String::new(),
-            1,
-            DeviceType::Network,
-            None,
-            true,
-        )
-        .save(&pool)
-        .await
-        .unwrap();
-        let device2 = Device::new(
-            "device2".to_string(),
-            String::new(),
-            1,
-            DeviceType::Network,
-            None,
-            true,
-        )
-        .save(&pool)
-        .await
-        .unwrap();
-
-        // device1 allowed
-        let _rd = AclRuleDevice {
-            id: NoId,
-            rule_id: rule.id,
-            device_id: device1.id,
-            allow: true,
-        }
-        .save(&pool)
-        .await
-        .unwrap();
-
-        // device2 denied
-        let _rd = AclRuleDevice {
-            id: NoId,
-            rule_id: rule.id,
-            device_id: device2.id,
-            allow: false,
-        }
-        .save(&pool)
-        .await
-        .unwrap();
-
-        // create 2 aliases
-        let alias1 = AclAlias::new(
-            "alias1",
-            AliasState::Applied,
-            AliasKind::Destination,
-            Vec::new(),
-            Vec::new(),
-            Vec::new(),
-        )
-        .save(&pool)
-        .await
-        .unwrap();
-        let _alias2 = AclAlias::new(
-            "alias2",
-            AliasState::Applied,
-            AliasKind::Destination,
-            Vec::new(),
-            Vec::new(),
-            Vec::new(),
-        )
-        .save(&pool)
-        .await
-        .unwrap();
-
-        // only alias1 applies to the rule
-        let _ra = AclRuleAlias {
-            id: NoId,
-            rule_id: rule.id,
-            alias_id: alias1.id,
-        }
-        .save(&pool)
-        .await
-        .unwrap();
-
-        let mut conn = pool.acquire().await.unwrap();
-
-        // convert to [`AclRuleInfo`] and verify results
-        let info = rule.to_info(&mut conn).await.unwrap();
-
-        assert_eq!(info.aliases.len(), 1);
-        assert_eq!(info.aliases[0], alias1);
-
-        assert_eq!(info.allowed_users.len(), 1);
-        assert_eq!(info.allowed_users[0], user1);
-
-        assert_eq!(info.denied_users.len(), 1);
-        assert_eq!(info.denied_users[0], user2);
-
-        assert_eq!(info.allowed_groups.len(), 1);
-        assert_eq!(info.allowed_groups[0], group1);
-
-        assert_eq!(info.denied_groups.len(), 1);
-        assert_eq!(info.denied_groups[0], group2);
-
-        assert_eq!(info.allowed_devices.len(), 1);
-        assert_eq!(info.allowed_devices[0].id, device1.id); // db modifies datetime precision
-
-        assert_eq!(info.denied_devices.len(), 1);
-        assert_eq!(info.denied_devices[0].id, device2.id); // db modifies datetime precision
-
-        assert_eq!(info.networks.len(), 1);
-        assert_eq!(info.networks[0], network1);
-
-        // test all_networks flag
-        rule.all_networks = true;
-        rule.save(&pool).await.unwrap();
-        assert_eq!(rule.get_networks(&pool).await.unwrap().len(), 2);
-
-        // test allowed/denied users
-        let allowed_users = rule.get_users(&pool, true).await.unwrap();
-        let denied_users = rule.get_users(&pool, false).await.unwrap();
-        assert_eq!(allowed_users.len(), 1);
-        assert_eq!(allowed_users[0], user1);
-        assert_eq!(denied_users.len(), 1);
-        assert_eq!(denied_users[0], user2);
-
-        // test `allow_all_users` flag
-        rule.allow_all_users = true;
-        rule.deny_all_users = false;
-        rule.save(&pool).await.unwrap();
-        assert_eq!(rule.get_users(&pool, true).await.unwrap().len(), 1);
-        assert_eq!(rule.get_users(&pool, false).await.unwrap().len(), 1);
-
-        // test `deny_all_users` flag
-        rule.allow_all_users = false;
-        rule.deny_all_users = true;
-        rule.save(&pool).await.unwrap();
-        assert_eq!(rule.get_users(&pool, true).await.unwrap().len(), 1);
-        assert_eq!(rule.get_users(&pool, false).await.unwrap().len(), 1);
-
-        // test both flags
-        rule.allow_all_users = true;
-        rule.deny_all_users = true;
-        rule.save(&pool).await.unwrap();
-        assert_eq!(rule.get_users(&pool, true).await.unwrap().len(), 1);
-        assert_eq!(rule.get_users(&pool, false).await.unwrap().len(), 1);
-
-        // deactivate user1
-        user1.is_active = false;
-        user1.save(&pool).await.unwrap();
-
-        // ensure only active users are allowed when `allow_all_users = true`
-        rule.allow_all_users = true;
-        rule.deny_all_users = false;
-        rule.save(&pool).await.unwrap();
-
-        let allowed_users = rule.get_users(&pool, true).await.unwrap();
-        let denied_users = rule.get_users(&pool, false).await.unwrap();
-        assert_eq!(allowed_users.len(), 0);
-        assert_eq!(denied_users.len(), 1);
-
-        // ensure only active users are allowed when `allow_all_users = false`
-        rule.allow_all_users = false;
-        rule.deny_all_users = false;
-        rule.save(&pool).await.unwrap();
-        ru2.allow = true; // allow user2
-        ru2.save(&pool).await.unwrap();
-        let allowed_users = rule.get_users(&pool, true).await.unwrap();
-        let denied_users = rule.get_users(&pool, false).await.unwrap();
-        assert_eq!(allowed_users.len(), 1);
-        assert_eq!(allowed_users[0], user2);
-        assert_eq!(denied_users.len(), 0);
-
-        // ensure only active users are denied when `deny_all_users = true`
-        rule.allow_all_users = false;
-        rule.deny_all_users = true;
-        rule.save(&pool).await.unwrap();
-
-        let allowed_users = rule.get_users(&pool, true).await.unwrap();
-        let denied_users = rule.get_users(&pool, false).await.unwrap();
-        assert_eq!(allowed_users.len(), 1);
-        assert_eq!(denied_users.len(), 0);
-
-        // ensure only active users are denied when `deny_all_users = false`
-        rule.allow_all_users = false;
-        rule.deny_all_users = false;
-        rule.save(&pool).await.unwrap();
-        ru2.allow = false; // deny user2
-        ru2.save(&pool).await.unwrap();
-        let allowed_users = rule.get_users(&pool, true).await.unwrap();
-        let denied_users = rule.get_users(&pool, false).await.unwrap();
-        assert_eq!(allowed_users.len(), 0);
-        assert_eq!(denied_users.len(), 1);
-        assert_eq!(denied_users[0], user2);
-    }
-
-    #[sqlx::test]
-    async fn test_all_allowed_users(_: PgPoolOptions, options: PgConnectOptions) {
-        let pool = setup_pool(options).await;
-
-        let mut rng = thread_rng();
-
-        // Create test users
-        let user_1: User<NoId> = rng.gen();
-        let user_1 = user_1.save(&pool).await.unwrap();
-        let user_2: User<NoId> = rng.gen();
-        let user_2 = user_2.save(&pool).await.unwrap();
-        let user_3: User<NoId> = rng.gen();
-        let user_3 = user_3.save(&pool).await.unwrap();
-        // inactive user
-        let mut user_4: User<NoId> = rng.gen();
-        user_4.is_active = false;
-        let user_4 = user_4.save(&pool).await.unwrap();
-
-        // Create test groups
-        let group_1 = Group {
-            id: NoId,
-            name: "group_1".into(),
-            ..Default::default()
-        }
-        .save(&pool)
-        .await
-        .unwrap();
-        let group_2 = Group {
-            id: NoId,
-            name: "group_2".into(),
-            ..Default::default()
-        }
-        .save(&pool)
-        .await
-        .unwrap();
-
-        // Assign users to groups:
-        // Group 1: users 1,2
-        // Group 2: user 3,4
-        let group_assignments = vec![
-            (&group_1, vec![&user_1, &user_2]),
-            (&group_2, vec![&user_3, &user_4]),
-        ];
-
-        for (group, users) in group_assignments {
-            for user in users {
-                query!(
-                    "INSERT INTO group_user (user_id, group_id) VALUES ($1, $2)",
-                    user.id,
-                    group.id
-                )
-                .execute(&pool)
-                .await
-                .unwrap();
-            }
-        }
-
-        // Create ACL rule
-        let rule = AclRule {
-            id: NoId,
-            name: "test_rule".to_string(),
-            allow_all_users: false,
-            deny_all_users: false,
-            allow_all_network_devices: false,
-            deny_all_network_devices: false,
-            all_networks: false,
-            destination: Vec::new(),
-            ports: Vec::new(),
-            protocols: Vec::new(),
-            expires: None,
-            enabled: true,
-            parent_id: None,
-            state: RuleState::Applied,
-        }
-        .save(&pool)
-        .await
-        .unwrap();
-
-        // Allow user_1 explicitly and group_2
-        AclRuleUser {
-            id: NoId,
-            rule_id: rule.id,
-            user_id: user_1.id,
-            allow: true,
-        }
-        .save(&pool)
-        .await
-        .unwrap();
-
-        AclRuleGroup {
-            id: NoId,
-            rule_id: rule.id,
-            group_id: group_2.id,
-            allow: true,
-        }
-        .save(&pool)
-        .await
-        .unwrap();
-
-        // Get rule info
-        let mut conn = pool.acquire().await.unwrap();
-        let rule_info = rule.to_info(&mut conn).await.unwrap();
-        assert_eq!(rule_info.allowed_users.len(), 1);
-        assert_eq!(rule_info.allowed_groups.len(), 1);
-
-        // Get all allowed users
-        let allowed_users = rule_info.get_all_allowed_users(&pool).await.unwrap();
-
-        // Should contain user1 (explicit) and user3 (from group2), but not inactive user_4
-        assert_eq!(allowed_users.len(), 2);
-        assert!(allowed_users.iter().any(|u| u.id == user_1.id));
-        assert!(allowed_users.iter().any(|u| u.id == user_3.id));
-        assert!(!allowed_users.iter().any(|u| u.id == user_4.id));
-    }
-
-    #[sqlx::test]
-    async fn test_all_denied_users(_: PgPoolOptions, options: PgConnectOptions) {
-        let pool = setup_pool(options).await;
-
-        let mut rng = thread_rng();
-
-        // Create test users
-        let user_1: User<NoId> = rng.gen();
-        let user_1 = user_1.save(&pool).await.unwrap();
-        let user_2: User<NoId> = rng.gen();
-        let user_2 = user_2.save(&pool).await.unwrap();
-        let user_3: User<NoId> = rng.gen();
-        let user_3 = user_3.save(&pool).await.unwrap();
-        // inactive user
-        let mut user_4: User<NoId> = rng.gen();
-        user_4.is_active = false;
-        let user_4 = user_4.save(&pool).await.unwrap();
-
-        // Create test groups
-        let group_1 = Group {
-            id: NoId,
-            name: "group_1".into(),
-            ..Default::default()
-        }
-        .save(&pool)
-        .await
-        .unwrap();
-        let group_2 = Group {
-            id: NoId,
-            name: "group_2".into(),
-            ..Default::default()
-        }
-        .save(&pool)
-        .await
-        .unwrap();
-
-        // Assign users to groups:
-        // Group 1: users 2,3,4
-        // Group 2: user 1
-        let group_assignments = vec![
-            (&group_1, vec![&user_2, &user_3, &user_4]),
-            (&group_2, vec![&user_1]),
-        ];
-
-        for (group, users) in group_assignments {
-            for user in users {
-                query!(
-                    "INSERT INTO group_user (user_id, group_id) VALUES ($1, $2)",
-                    user.id,
-                    group.id
-                )
-                .execute(&pool)
-                .await
-                .unwrap();
-            }
-        }
-
-        // Create ACL rule
-        let rule = AclRule {
-            id: NoId,
-            name: "test_rule".to_string(),
-            allow_all_users: false,
-            deny_all_users: false,
-            allow_all_network_devices: false,
-            deny_all_network_devices: false,
-            all_networks: false,
-            destination: Vec::new(),
-            ports: Vec::new(),
-            protocols: Vec::new(),
-            expires: None,
-            enabled: true,
-            parent_id: None,
-            state: RuleState::Applied,
-        }
-        .save(&pool)
-        .await
-        .unwrap();
-
-        // Deny user_1, user_3 explicitly and group_1
-        AclRuleUser {
-            id: NoId,
-            rule_id: rule.id,
-            user_id: user_1.id,
-            allow: false,
-        }
-        .save(&pool)
-        .await
-        .unwrap();
-        AclRuleUser {
-            id: NoId,
-            rule_id: rule.id,
-            user_id: user_3.id,
-            allow: false,
-        }
-        .save(&pool)
-        .await
-        .unwrap();
-
-        AclRuleGroup {
-            id: NoId,
-            rule_id: rule.id,
-            group_id: group_1.id,
-            allow: false,
-        }
-        .save(&pool)
-        .await
-        .unwrap();
-
-        // Get rule info
-        let mut conn = pool.acquire().await.unwrap();
-        let rule_info = rule.to_info(&mut conn).await.unwrap();
-        assert_eq!(rule_info.denied_users.len(), 2);
-        assert_eq!(rule_info.denied_groups.len(), 1);
-
-        // Get all denied users
-        let denied_users = rule_info.get_all_denied_users(&pool).await.unwrap();
-
-        // Should contain user_1 (explicit), user_2 and user_3 (from group_1), but not inactive user_4
-        assert_eq!(denied_users.len(), 3);
-        assert!(denied_users.iter().any(|u| u.id == user_1.id));
-        assert!(denied_users.iter().any(|u| u.id == user_2.id));
-        assert!(denied_users.iter().any(|u| u.id == user_3.id));
-        assert!(!denied_users.iter().any(|u| u.id == user_4.id));
-    }
-}
+mod tests;
