+++ src/enterprise/mod.rs
@@ -6,6 +6,7 @@ pub mod handlers;
 pub mod ldap;
 pub mod license;
 pub mod limits;
+mod utils;
 
 use license::{get_cached_license, validate_license};
 use limits::get_counts;
